pyproject.toml
[project]
name = "tara"
version = "0.1.0"
description = "Agente que calcula porções ideais de alimentos baseado no seu perfil de saúde"
authors = [{name = "Yasmim Ferreira"}, {name = "Camille Martins"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.31.0",
    "g4f>=0.3.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "prisma>=0.15.0",
    "bcrypt>=4.2.0",
    "PyJWT>=2.9.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tara/api/tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
]
